Clean up work
In this project I'm attempting to make a basic implementation of IA-DEVS
To keep it simpling, in this commit I'm first removing several Cadmium
complexities that are not fundamental to the project.
- Removing cell-devs, there is no cell-devs equivalent yet in IA-DEVS
- Upgrading dependency requirements, C++ 23 minimum, we want to remove
  old work around to the lack of concepts implementation back in 2013.
- Removing unneeded dependencies.

In the next few commits, I'm going to transform old C++ code into
simplified version using new features of the language. After that,
I will start the real implementation.